back/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.39
psycopg[binary]==3.2.6
alembic==1.15.2
pydantic==2.11.1
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
python-multipart==0.0.20
